% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivetime.R
\name{calc_point_distance}
\alias{calc_point_distance}
\title{Find point certain duration away in a straight line}
\usage{
calc_point_distance(start, end, duration)
}
\arguments{
\item{start}{starting point in the form of a geocode list, index 1 is longitude, index 2 is latitude}

\item{end}{end point in the form of a geocode list, index 1 is longitude, index 2 is latitude}

\item{duration}{drive duration we are looking to find, in minutes}
}
\value{
geocode of the point that is approximately 'duration' minutes away
}
\description{
Function takes a starting point and end point to make a line, then finds
a point on that line that would take a certain amount of time to drive to.
}
\details{
Creates a list of geocodes along a line between the start and end point, and uses a binary search
algorithm to find the point that is closest to the correct distance.
}
\author{
Martin Boros
}
