% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivetime.R
\name{drive_time_points}
\alias{drive_time_points}
\title{Create dataframe of geocodes that can be used to map a drivetime isochrone}
\usage{
drive_time_points(lat, lng, duration, n = 60)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{duration}{driving distance}

\item{n}{number of points in the isochrone}
}
\value{
data frame of geocodes for isochrone to be used in conjunction with a mapping package
}
\description{
Functions takes a geocode and desired driving duration to create a drivetime isochrone.
}
\details{
Taking a geocode, this will create a drivetime isochrone. Fidelity increases as n increases, but
because the user has a limited number of calls to the distance matrix API daily and there are
roughly 3*n calls made to the API for each call of drive_time_points, one can quickly burn up
all of their daily calls by creating a few drivetime isochrones with a high n value.
Currently only supports minutes and miles, and the only valid input is a geocode.
}
\author{
Martin Boros
}
