% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Get lat / lng coordinates from the Google Maps Geocode API.}
\usage{
geocode(address, key = "AIzaSyCp3l9A5PKPVupslJqOFmO55kg50JYd85Q")
}
\arguments{
\item{address}{An address string}

\item{key}{A rgeistered Google Maps Geocode API key.}
}
\value{
A set of latitude / longitude coordinates associated with a requested address.
}
\description{
Get lat / lng coordinates from the Google Maps Geocode API.
}
