% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivetime.R
\name{get_distance}
\alias{get_distance}
\title{Getting drive distance using Google Distance Matrix API}
\usage{
get_distance(from_lat, from_lng, to_lat, to_lng)
}
\arguments{
\item{from_lat}{is the starting latitude}

\item{from_lng}{is the starting longitude}

\item{to_lat}{is the ending latitude}

\item{to_lng}{is the ending longitude}
}
\value{
data frame that contains the distance matrix data from Google's API
}
\description{
This function gets the distance matrix between two geocodes
}
\details{
The function calls the distance matrix API from Google which returns a JSON containing the destination address,
the origin address, the distance in miles (character), distance in meters (numeric), the duration in hours and minutes(character),
the distance in minutes(numeric), and the call STATUS.
Can be used the same way as the 'mapdist' function in ggmap, but works better with geocodes since it does not require a reverse geocode
which often times fails.
}
\author{
Martin Boros
}
