% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{rev.geocode}
\alias{rev.geocode}
\title{Get address from the Google Maps Geocode API.}
\usage{
\method{rev}{geocode}(lat, lng,
  key = "AIzaSyCp3l9A5PKPVupslJqOFmO55kg50JYd85Q")
}
\arguments{
\item{lat}{A latitude coordinate}

\item{lng}{A longitude coordinate}

\item{key}{A rgeistered Google Maps Geocode API key.}
}
\value{
An address associated with a requested set of location coordinates.
}
\description{
Get address from the Google Maps Geocode API.
}
